Issue #3403634 by SV: Remove redundant group relation type for nodes; Refactor access control for events
