+++ modules/social_features/social_event/modules/social_event_managers/social_event_managers.services.yml
@@ -14,3 +14,8 @@ services:
     class: 'Drupal\social_event_managers\Access\AddEnrolleeAccessCheck'
     tags:
       - { name: 'access_check', applies_to: '_enrollee_permission', priority: 249 }
+
+  group.relation_handler.access_control.group_node:
+    class: 'Drupal\social_event_managers\Plugin\Group\RelationHandlerDefault\EventsGroupContentAccessControl'
+    arguments: ['@group.relation_handler.access_control']
+    shared: false
