+++ modules/social_features/social_group/modules/social_group_request/social_group_request.install
@@ -38,10 +38,9 @@ function social_group_request_update_dependencies(): array {
  */
 function _social_group_request_set_permissions() {
   if (\Drupal::moduleHandler()->moduleExists('social_group_flexible_group')) {
-    $group_type = GroupType::load('flexible_group');
     $outsider = \Drupal::entityTypeManager()
       ->getStorage('group_role')
-      ->load($group_type->id() . '-outsider');
+      ->load('flexible_group-outsider');
     $outsider->grantPermissions(['request group membership'])->save();
   }
 }
